+++ code/solver/include/solver/algorithms/wcrossover/add.hpp
@@ -29,10 +29,10 @@ namespace uscp::wcrossover
 		add& operator=(const add& other) = delete;
 		add& operator=(add&& other) noexcept = delete;
 
-		[[nodiscard]] std::vector<ssize_t> apply(const std::vector<ssize_t>& a,
-		                                         const std::vector<ssize_t>& b) const noexcept
+		[[nodiscard]] std::vector<long long> apply(const std::vector<long long>& a,
+		                                           const std::vector<long long>& b) const noexcept
 		{
-			std::vector<ssize_t> weights;
+			std::vector<long long> weights;
 			weights.reserve(problem.points_number);
 			std::transform(std::cbegin(a),
 			               std::cend(a),
@@ -42,17 +42,17 @@ namespace uscp::wcrossover
 			return weights;
 		}
 
-		std::vector<ssize_t> apply1(const std::vector<ssize_t>& a,
-		                            const std::vector<ssize_t>& b,
-		                            [[maybe_unused]] random_engine& generator) const noexcept
+		std::vector<long long> apply1(const std::vector<long long>& a,
+		                              const std::vector<long long>& b,
+		                              [[maybe_unused]] random_engine& generator) const noexcept
 		{
 			assert(a.size() == problem.points_number && b.size() == problem.points_number);
 			return apply(a, b);
 		}
 
-		std::vector<ssize_t> apply2(const std::vector<ssize_t>& a,
-		                            const std::vector<ssize_t>& b,
-		                            [[maybe_unused]] random_engine& generator) const noexcept
+		std::vector<long long> apply2(const std::vector<long long>& a,
+		                              const std::vector<long long>& b,
+		                              [[maybe_unused]] random_engine& generator) const noexcept
 		{
 			assert(a.size() == problem.points_number && b.size() == problem.points_number);
 			return apply(a, b);
