+++ code/solver/include/solver/algorithms/wcrossover/difference.hpp
@@ -29,30 +29,30 @@ namespace uscp::wcrossover
 		difference& operator=(const difference& other) = delete;
 		difference& operator=(difference&& other) noexcept = delete;
 
-		[[nodiscard]] std::vector<ssize_t> apply(const std::vector<ssize_t>& a,
-		                                         const std::vector<ssize_t>& b) const noexcept
+		[[nodiscard]] std::vector<long long> apply(const std::vector<long long>& a,
+		                                           const std::vector<long long>& b) const noexcept
 		{
-			std::vector<ssize_t> weights;
+			std::vector<long long> weights;
 			weights.reserve(problem.points_number);
 			std::transform(std::cbegin(a),
 			               std::cend(a),
 			               std::cbegin(b),
 			               std::back_inserter(weights),
-			               [](ssize_t wa, ssize_t wb) { return wa > wb ? wa - wb : wb - wa; });
+			               [](long long wa, long long wb) { return wa > wb ? wa - wb : wb - wa; });
 			return weights;
 		}
 
-		std::vector<ssize_t> apply1(const std::vector<ssize_t>& a,
-		                            const std::vector<ssize_t>& b,
-		                            [[maybe_unused]] random_engine& generator) const noexcept
+		std::vector<long long> apply1(const std::vector<long long>& a,
+		                              const std::vector<long long>& b,
+		                              [[maybe_unused]] random_engine& generator) const noexcept
 		{
 			assert(a.size() == problem.points_number && b.size() == problem.points_number);
 			return apply(a, b);
 		}
 
-		std::vector<ssize_t> apply2(const std::vector<ssize_t>& a,
-		                            const std::vector<ssize_t>& b,
-		                            [[maybe_unused]] random_engine& generator) const noexcept
+		std::vector<long long> apply2(const std::vector<long long>& a,
+		                              const std::vector<long long>& b,
+		                              [[maybe_unused]] random_engine& generator) const noexcept
 		{
 			assert(a.size() == problem.points_number && b.size() == problem.points_number);
 			return apply(a, b);
