+++ code/solver/include/solver/algorithms/wcrossover/minmax.hpp
@@ -29,33 +29,33 @@ namespace uscp::wcrossover
 		minmax& operator=(const minmax& other) = delete;
 		minmax& operator=(minmax&& other) noexcept = delete;
 
-		std::vector<ssize_t> apply1(const std::vector<ssize_t>& a,
-		                            const std::vector<ssize_t>& b,
-		                            [[maybe_unused]] random_engine& generator) const noexcept
+		std::vector<long long> apply1(const std::vector<long long>& a,
+		                              const std::vector<long long>& b,
+		                              [[maybe_unused]] random_engine& generator) const noexcept
 		{
 			assert(a.size() == problem.points_number && b.size() == problem.points_number);
-			std::vector<ssize_t> weights;
+			std::vector<long long> weights;
 			weights.reserve(problem.points_number);
 			std::transform(std::cbegin(a),
 			               std::cend(a),
 			               std::cbegin(b),
 			               std::back_inserter(weights),
-			               [](ssize_t wa, ssize_t wb) { return wa > wb ? wa : wb; });
+			               [](long long wa, long long wb) { return wa > wb ? wa : wb; });
 			return weights;
 		}
 
-		std::vector<ssize_t> apply2(const std::vector<ssize_t>& a,
-		                            const std::vector<ssize_t>& b,
-		                            [[maybe_unused]] random_engine& generator) const noexcept
+		std::vector<long long> apply2(const std::vector<long long>& a,
+		                              const std::vector<long long>& b,
+		                              [[maybe_unused]] random_engine& generator) const noexcept
 		{
 			assert(a.size() == problem.points_number && b.size() == problem.points_number);
-			std::vector<ssize_t> weights;
+			std::vector<long long> weights;
 			weights.reserve(problem.points_number);
 			std::transform(std::cbegin(a),
 			               std::cend(a),
 			               std::cbegin(b),
 			               std::back_inserter(weights),
-			               [](ssize_t wa, ssize_t wb) { return wa < wb ? wa : wb; });
+			               [](long long wa, long long wb) { return wa < wb ? wa : wb; });
 			return weights;
 		}
